better error

+++ scripts/workman.fish
@@ -5,6 +5,15 @@
 set script_name (basename (status --current-filename))
 set script_path (dirname (realpath (status --current-filename)))
 set script_version "1.0.0"
+set dry_run false
+
+function execute
+    if $dry_run
+        echo $argv
+    else
+        $argv
+    end
+end
 
 function show_help
     echo "Usage: $script_name <command> [options]"
@@ -29,27 +38,41 @@ end
 
 function init_image
     set -l project_root (git rev-parse --show-toplevel)
+    if test -z "$project_root"
+        echo "✗ Not a git repository, please `git init` first" >&2
+        return 1
+    end
 
     echo "Creating Dockerfile in $project_root" >&2
-    cp $script_path/Dockerfile $project_root/Dockerfile
+    execute cp $script_path/Dockerfile $project_root/Dockerfile
 
     echo "Adding .worktrees to .gitignore" >&2
-    mkdir -p $project_root/.worktrees/
-    grep -q '^/.worktrees/$' $project_root/.gitignore 2> /dev/null; or echo /.worktrees/ >> $project_root/.gitignore
+    execute mkdir -p $project_root/.worktrees/
+    grep -q '^/.worktrees/$' $project_root/.gitignore 2> /dev/null; or execute fish -c "echo /.worktrees/ >> $project_root/.gitignore"
 end
 
 function get_project_path
-    git worktree list --porcelain -z | awk -v RS='\0' '/^worktree / {print substr($0, 10)}' | head
+    git worktree list --porcelain -z | awk -v RS='\0' '/^worktree / {print substr($0, 10)}' | head -1
 end
 
 function build_image
     set -l project_root (get_project_path)
+    if test -z "$project_root"
+        echo "✗ Not a git repository, please `git init` first" >&2
+        return 1
+    end
     set -l project_name (basename $project_root)
+
+    if test -z "$project_name"; or not test -f Dockerfile
+        echo "✗ Dockerfile not found, please `$script_name build` first" >&2
+        return 1
+    end
+
     # cp ~/.config/opencode/opencode.json .worktrees/opencode.json
-    docker build \
+    execute docker build \
         --build-arg "HTTP_PROXY=$(string replace 127.0.0.1 172.17.0.1 $http_proxy)" \
         --build-arg "HTTPS_PROXY=$(string replace 127.0.0.1 172.17.0.1 $https_proxy)" \
-        -t git-worker-for-$project_name:latest $project_root
+        -t "git-worker-for-$project_name:latest" $project_root
     # rm -f .worktrees/opencode.json
 end
 
@@ -61,16 +84,32 @@ function run_worktree --argument-names branch
     end
 
     set -l project_root (get_project_path)
+    if test -z "$project_root"
+        echo "✗ Not a git repository, please `git init` first" >&2
+        return 1
+    end
     set -l project_name (basename $project_root)
-    docker run \
+
+    set -l worktree_path (get_worktree_path $branch)
+    if test $status -ne 0
+        echo "✗ Branch or worktree not found" >&2
+        return 1
+    end
+
+    if test "$(docker images --format json | grep "git-worker-for-$project_name" | jq .Repository -r)" != "git-worker-for-$project_name"
+        echo "✗ Image not found, please `$script_name build` first" >&2
+        return 1
+    end
+
+    execute docker run \
         -e "http_proxy=$(string replace 127.0.0.1 172.17.0.1 $http_proxy)" \
         -e "https_proxy=$(string replace 127.0.0.1 172.17.0.1 $https_proxy)" \
         -e "all_proxy=$(string replace 127.0.0.1 172.17.0.1 $all_proxy)" \
         -e "no_proxy=$no_proxy,172.17.0.1" \
-        -w "$project_root/.worktrees/$branch/$project_name" \
+        -w "$worktree_path" \
         -v "$HOME/.config/opencode:/home/ubuntu/.config/opencode:ro" \
         -v "$project_root:$project_root:ro" \
-        -v "$project_root/.worktrees/$branch/$project_name:$project_root/.worktrees/$branch/$project_name" \
+        -v "$worktree_path:$worktree_path" \
         -u (id -u):(id -g) \
         --add-host=host.docker.internal:host-gateway \
         --rm -it git-worker-for-$project_name:latest $argv[2..]
@@ -81,13 +120,9 @@ function list_worktrees
 end
 
 function get_worktree_path --argument-names branch
-    git check-ref-format --branch "$branch" >/dev/null || return 1
-    git worktree list --porcelain \
-        | grep  "^branch refs/heads/$branch" >/dev/null \
-        || return 1
-    git worktree list --porcelain \
-        | grep -B2 "^branch refs/heads/$branchName" \
-        | head -1 | cut -d' ' -f 2
+    git check-ref-format --branch "$branch" >/dev/null; or return 1
+    git worktree list --porcelain | grep  "^branch refs/heads/$branch" >/dev/null ; or return 1
+    git worktree list --porcelain | grep -B2 "^branch refs/heads/$branch" | head -1 | cut -d' ' -f 2
 end
 
 function path_of_worktree --argument-names branch
@@ -130,10 +165,15 @@ function add_worktree --argument-names branch
     end
 
     set -l project_root (get_project_path)
+    if test -z "$project_root"
+        echo "✗ Not a git repository, please `git init` first" >&2
+        return 1
+    end
     set -l project_name (basename $project_root)
+
     set -l worktree_path "$project_root/.worktrees/$branch/$project_name"
 
-    echo "Adding worktree for branch: $branch" >&2
+    echo "Adding worktree for branch: $branch [$worktree_path]" >&2
 
     # Create new worktree
     git worktree add -b "$branch" "$worktree_path"
@@ -154,13 +194,17 @@ function remove_worktree --argument-names branch
     end
 
     set -l project_root (get_project_path)
+    if test -z "$project_root"
+        echo "✗ Not a git repository, please `git init` first" >&2
+        return 1
+    end
     set -l project_name (basename $project_root)
     set -l worktree_path "$project_root/.worktrees/$branch/$project_name"
 
     echo "Removing worktree for branch: $branch" >&2
 
     # Remove worktree
-    git worktree remove "$worktree_path"
+    execute git worktree remove "$worktree_path"
 
     if test $status -ne 0
         echo "✗ Failed to remove worktree" >&2
@@ -168,7 +212,7 @@ function remove_worktree --argument-names branch
     end
 
     # Delete branch (silently)
-    git branch -dq "$branch"
+    execute git branch -dq "$branch"
 
     if test $status -ne 0
         echo "✗ Failed to delete branch" >&2
